обернул все. оставил только load
